+++ internal/image/install/install.go
@@ -20,7 +20,7 @@ package install
 import (
 	"regexp"
 
-	"github.com/suse/elemental/v3/pkg/security"
+	"github.com/suse/elemental/v3/pkg/crypto"
 )
 
 type DiskSize string
@@ -30,8 +30,8 @@ func (d DiskSize) IsValid() bool {
 }
 
 type Installation struct {
-	Bootloader     string          `yaml:"bootloader"`
-	KernelCmdLine  string          `yaml:"kernelCmdLine"`
-	DiskSize       DiskSize        `yaml:"diskSize"`
-	SecurityPolicy security.Policy `yaml:"securityPolicy"`
+	Bootloader    string        `yaml:"bootloader"`
+	KernelCmdLine string        `yaml:"kernelCmdLine"`
+	DiskSize      DiskSize      `yaml:"diskSize"`
+	CryptoPolicy  crypto.Policy `yaml:"cryptoPolicy"`
 }
