+++ pkg/crypto/policy.go
@@ -15,7 +15,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package security
+package crypto
 
 const (
 	DefaultPolicy Policy = "default"
